Change AttributeScanner to Action::ComparesAttributes
